Move search route & template to theme
