+++ packages/global/routes/index.js
@@ -1,12 +1,12 @@
 const htmlSitemap = require('@parameter1/base-cms-marko-web-html-sitemap/routes');
 const renderBlock = require('@randall-reilly/package-theme-monorail/routes/render-block');
+const search = require('@randall-reilly/package-theme-monorail/routes/search');
 const feed = require('./feed');
 const nativeX = require('./native-x');
 const omedaNewsletters = require('./omeda-newsletters');
 const printContent = require('./print-content');
 const publicFiles = require('./public-files');
 const redirects = require('./redirects');
-const search = require('./search');
 const staticPage = require('./static-page');
 const taxonomy = require('./taxonomy');
