+++ packages/theme-monorail/routes/search.js
