+++ packages/theme-monorail/templates/search.marko
